api/requirements.txt
# Django & DRF
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
djoser>=2.2
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27

# Database
psycopg2-binary>=2.9

# Utils
python-dotenv>=1.0
gunicorn>=21.0

# Dev tools
black>=24.0
ruff>=0.4
mypy>=1.10
django-stubs>=5.0
djangorestframework-stubs>=3.14
